+++ ARKit-CubiCasa/Resources/Assets.xcassets/AppIcon.appiconset/Contents.json
