+++ ARKit-CubiCasa/Resources/Assets.xcassets/DropObjectIcon.imageset/Contents.json
