+++ ARKit-CubiCasa/Resources/Assets.xcassets/DropObjectIcon.imageset/image@1x.png
