+++ ARKit-CubiCasa/Resources/Assets.xcassets/DropObjectIcon.imageset/image@2x.png
