+++ ARKit-CubiCasa/Resources/Assets.xcassets/DropObjectIcon.imageset/image@3x.png
